backend/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.0
psycopg2-binary==2.9.9
PyJWT==2.10.1
djangorestframework-simplejwt==5.5.1
django-filter==23.4
python-decouple==3.8
openai==1.3.5
python-dotenv==1.0.0
gunicorn==21.2.0
whitenoise==6.6.0
pillow==10.1.0
django-storages==1.14.2
boto3==1.29.7
requests==2.31.0
celery==5.3.4
redis==5.0.1
